Integrating authentication, fixed missing packages, Makefile typo fix
