--- 42-auth-backend/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "singleQuote": true,
-  "trailingComma": "all"
-}
