--- 42-auth-backend/src/main.ts
@@ -1,11 +1,0 @@
-// src/main.ts
-
-import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
-
-async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-  app.enableCors();
-  await app.listen(3001);
-}
-bootstrap();
