+++ frontend/package.json
@@ -14,10 +14,12 @@
   "dependencies": {
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "react-router-dom": "^6.26.2",
     "react-scripts": "5.0.0",
     "socket.io-client": "^4.0.0"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.20.7",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
@@ -26,8 +28,7 @@
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.0",
     "prettier": "^3.0.0",
-    "typescript": "^4.9.5",
-    "@babel/plugin-proposal-private-property-in-object": "^7.20.7"
+    "typescript": "^4.9.5"
   },
   "browserslist": {
     "production": [
